+++ app/src/main/java/eu/kanade/tachiyomi/data/download/DownloadProvider.kt
@@ -6,7 +6,6 @@ import eu.kanade.tachiyomi.source.Source
 import eu.kanade.tachiyomi.util.storage.DiskUtil
 import logcat.LogPriority
 import tachiyomi.core.common.i18n.stringResource
-import tachiyomi.core.common.storage.displayablePath
 import tachiyomi.core.common.util.system.logcat
 import tachiyomi.domain.chapter.model.Chapter
 import tachiyomi.domain.manga.model.Manga
@@ -14,6 +13,7 @@ import tachiyomi.domain.storage.service.StorageManager
 import tachiyomi.i18n.MR
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
+import java.io.IOException
 
 /**
  * This class is used to provide the directories where the downloads should be saved.
@@ -36,18 +36,21 @@ class DownloadProvider(
      * @param source the source of the manga.
      */
     internal fun getMangaDir(mangaTitle: String, source: Source): UniFile {
+        val sourceDirName = getSourceDirName(source)
+        val mangaDirName = getMangaDirName(mangaTitle)
         try {
-            return downloadsDir!!
-                .createDirectory(getSourceDirName(source))!!
-                .createDirectory(getMangaDirName(mangaTitle))!!
+            val downloadsDir = downloadsDir
+            if (downloadsDir == null) {
+                logcat(LogPriority.ERROR) { "Invalid download directory or No permission allowed" }
+                throw SecurityException(context.stringResource(MR.strings.invalid_download_directory))
+            }
+            return downloadsDir.createDirectory(sourceDirName)!!
+                .createDirectory(getMangaDirName(mangaDirName))!!
         } catch (e: Throwable) {
-            logcat(LogPriority.ERROR, e) { "Invalid download directory" }
-            throw Exception(
-                context.stringResource(
-                    MR.strings.invalid_location,
-                    downloadsDir?.displayablePath ?: "",
-                ),
-            )
+            if (e is SecurityException) throw e
+            val mangaDir = "/$sourceDirName/$mangaDirName"
+            logcat(LogPriority.ERROR, e) { "Invalid location: $mangaDir" }
+            throw IOException(context.stringResource(MR.strings.invalid_location, mangaDir))
         }
     }